fixed a performance counter bug...but another one resides there...what if doubles lose precision

+++ src/lin_time.cpp
@@ -80,7 +80,7 @@ double GetElapsedTime() {
 void UpdateTime() {
 #ifdef WIN32
   QueryPerformanceCounter((LARGE_INTEGER*)&newtime);
-  elapsedtime = ((float)(newtime-ttime))/freq;
+  elapsedtime = ((double)(newtime-ttime))/freq;
   ttime = newtime;
 #else
   lasttime = newtime;
